updated tip tracking in all protocols

+++ protocols/S3/stationA_protocol2_beads_S3.ot2.apiv2.py
@@ -82,6 +82,7 @@ def finish_run():
 def retrieve_tip_info(pip,tipracks,file_path = '/data/A/tip_log.json'):
     global tip_log
     if not tip_log['count'] or pip not in tip_log['count']:
+        tip_log['count'][pip] = 0
         if not robot.is_simulating():
             if os.path.isfile(file_path):
                 with open(file_path) as json_file:
@@ -90,12 +91,8 @@ def retrieve_tip_info(pip,tipracks,file_path = '/data/A/tip_log.json'):
                         tip_log['count'][pip] = data['tips1000']
                     elif 'P300' in str(pip):
                         tip_log['count'][pip] = data['tips300']
-                    else:
-                        tip_log['count'][pip] = 0
-            else:
-                tip_log['count'][pip] = 0
-        else:
-            tip_log['count'][pip] = 0
+                    elif 'P20' in str(pip):
+                        tip_log['count'][pip] = data['tips20']
 
         if "8-Channel" in str(pip):
             tip_log['tips'][pip] =  [tip for rack in tipracks for tip in rack.rows()[0]]
@@ -109,12 +106,15 @@ def retrieve_tip_info(pip,tipracks,file_path = '/data/A/tip_log.json'):
 def save_tip_info(file_path = '/data/A/tip_log.json'):
     data = {}
     if not robot.is_simulating():
-        os.rename(file_path,file_path + ".bak")
+        if os.path.isfile(file_path):
+            os.rename(file_path,file_path + ".bak")
         for pip in tip_log['count']:
             if "P1000" in str(pip):
                 data['tips1000'] = tip_log['count'][pip]
             elif "P300" in str(pip):
                 data['tips300'] = tip_log['count'][pip]
+            elif "P20" in str(pip):
+                data['tips20'] = tip_log['count'][pip]
 
         with open(file_path, 'a+') as outfile:
             json.dump(data, outfile)
@@ -135,10 +135,14 @@ resuming.')
 
 def drop(pip):
     global switch
-    side = 1 if switch else -1
-    drop_loc = robot.loaded_labwares[12].wells()[0].top().move(Point(x=side*20))
-    pip.drop_tip(drop_loc,home_after=False)
-    switch = not switch
+    if "8-Channel" not in str(pip):
+        side = 1 if switch else -1
+        drop_loc = robot.loaded_labwares[12].wells()[0].top().move(Point(x=side*20))
+        pip.drop_tip(drop_loc,home_after=False)
+        switch = not switch
+    else:
+        drop_loc = robot.loaded_labwares[12].wells()[0].top().move(Point(x=20))
+        pip.drop_tip(drop_loc,home_after=False)
 
 def prepare_beads(bd_tube,eth_tubes,pip,tiprack):
     pick_up(pip,tiprack)
